+++ common/src/main/java/pepjebs/mapatlases/config/MapAtlasesConfig.java
@@ -11,7 +11,7 @@ import java.util.function.Supplier;
 public class MapAtlasesConfig {
 
     static {
-        ConfigBuilder builder = ConfigBuilder.create(MapAtlasesMod.MOD_ID, ConfigType.COMMON);
+        ConfigBuilder builder = ConfigBuilder.create(MapAtlasesMod.MOD_ID, ConfigType.COMMON_SYNCED);
 
 
         builder.push("general");
@@ -54,6 +54,9 @@ public class MapAtlasesConfig {
         lightMap = builder.comment("Shows light color on maps. Needs Moonlight lib")
                 .define("light_map", false);
 
+        entityRadar = builder.comment("Show nearby mobs on minimap. Needs matching client config also set")
+                .define("mob_radar", false);
+
         shearButton = builder.comment("Adds a shear button to the atlas screen which allows you to cut maps")
                 .define("shear_button", true);
 
@@ -90,6 +93,7 @@ public class MapAtlasesConfig {
     public static final Supplier<Boolean> creativeTeleport;
     public static final Supplier<Boolean> roundRobinUpdate;
     public static final Supplier<Boolean> lightMap;
+    public static final Supplier<Boolean> entityRadar;
     public static final Supplier<String> pinMarkerId;
     public static final Supplier<Boolean> shearButton;
     public static final Supplier<Integer> mapUpdatePerTick;